[CSJHW-5]: task done

+++ src/Route256.Week5.Homework.PriceCalculator.Api/Requests/V1/ClearHistoryRequest.cs
@@ -1,3 +1,3 @@
 ﻿namespace Route256.Week5.Homework.PriceCalculator.Api.Requests.V1;
 
-public record ClearHistoryRequest(int UserId, int[] CalculationIds);
+public record ClearHistoryRequest(long UserId, long[] CalculationIds);
